requirements.txt
disnake
requests
python-dotenv
boto3
flask
cryptography
asyncpg
aiohttp
brotli